improvement on test
